% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixImagesForOcto}
\alias{fixImagesForOcto}
\title{Get list of images in document}
\usage{
fixImagesForOcto(filename)
}
\description{
returns lines--does not write them.
Side effects: copies images to folder specified by options("imagedir")
}

