% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit2wiki.R
\name{knit2wiki}
\alias{knit2wiki}
\title{Knit .Rmd files to personal wiki directory as properly formatted .html}
\usage{
knit2wiki(file, generate = FALSE, format = "html_document",
  subdir = "inbox", wikidir = getOption("wikidir"), ...)
}
\arguments{
\item{file}{The .Rmd file to knit}

\item{generate}{`rake generate` the website? Defaults to FALSE.}

\item{format}{How to render the document. See ?rmarkdown::render for options}

\item{subdir}{Location to put file relative to wikidir. Defaults to "inbox"}

\item{wikidir}{Location of wiki library (assuming Wikitten structure)}

\item{...}{passed to rmarkdown::render}
}
\description{
This function knits a .Rmd file to .html and sends the result to a personal
wiki source directory with proper yaml front matter
}

