% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderWithWidgets.R
\name{renderWithWidgets}
\alias{renderWithWidgets}
\title{Render function for htmlwidgets}
\usage{
renderWithWidgets(input_file, output_file = sub("\\\\.Rmd$", ".html",
  input_file, ignore.case = TRUE), self_contained = TRUE,
  deps_path = file.path(dirname(output_file), "deps"))
}
\arguments{
\item{input_file}{input file}

\item{output_file}{output file}

\item{self_contained}{Should the output be a self-contained html file?}

\item{deps_path}{path of deps.}
}
\description{
Stolen from http://stackoverflow.com/questions/31645528/knit-dtdatatable-without-pandoc
I don't know why this works, or even the nature of the problem it solves
}

